# repo/api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: KuberCloud ANI Core API
  version: 1.0.0
  description: |
    KuberCloud ANI Core — 基础设施平台 REST API 契约（唯一真实来源）。

    ## 文件范围与边界
    本文件（api/openapi/v1.yaml）仅包含 ANI Core 基础设施资源：
    instances（VM/Container/GPU/Sandbox/BM/K8s集群）、volumes/filesystems/objects/
    vector-stores、networks、auth/secrets/encryption、gpu-inventory/dpu-inventory、
    registry/metering/observability/notifications/audit 等。
    ANI Services 资源（模型仓库/推理/知识库/PaaS）在 api/openapi/services/v1.yaml。

    ## URL 前缀规范
    servers[0].url = "https://{host}/api/v1"（前缀在 spec 中定义，Gateway 不做路径变换）。
    版本升级时新建 api/openapi/v2.yaml，本文件路径保持稳定。

    ## SDK 生成（make gen-core-sdk）
    Go: github.com/kubercloud/ani-go | Python: pip install kubercloud-ani
    TypeScript: @kubercloud/ani-client | Java: com.kubercloud:ani-java

    ## 认证
    Bearer JWT（1小时有效）或 X-API-Key Header。
    tenant_id 从 JWT claims 提取，请求体中的 tenant_id 字段将被忽略。

    ## 错误格式
    `{ code: string, message: string, request_id: string, details?: object }`
    标准错误码：UNAUTHORIZED / FORBIDDEN / NOT_FOUND / CONFLICT / BAD_REQUEST /
    RATE_LIMIT_EXCEEDED / NOT_IMPLEMENTED / INTERNAL_ERROR

    ## 分页
    cursor 分页：`?limit=20&cursor=<token>` → `{ items: [], next_cursor: string | null }`

    ## 异步操作
    202 响应体返回 AsyncTask，Location Header 附任务 URL。
    轮询 GET /api/v1/tasks/{task_id} 或配置 Webhook。
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: ani.internal
        description: ANI Gateway 地址（支持IP地址）

...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT AccessToken（有效期1小时，通过 /auth/token 获取）
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 长期 API Key（在 Console → 设置 → API Key 中创建）

# ── 公共 Schema ───────────────────────────────────────────────────────────────
  schemas:
...
    EncryptionKeyCreateRequest:
      type: object
      properties:
        idempotency_key: { type: string }
        name:            { type: string }
        algorithm:       { type: string, enum: [SM4, AES256] }
      required: [idempotency_key, name]
    EncryptionKey:
      type: object
      properties:
        id:         { type: string }
        tenant_id:  { type: string }
        name:       { type: string }
        algorithm:  { type: string }
        state:      { type: string, enum: [active, rotated, revoked, deleted] }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    EncryptionKeyListResponse:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/EncryptionKey" } }
        total: { type: integer }
        next_cursor: { type: string, nullable: true }
    EncryptionKeyRotateRequest:
      type: object
      properties:
        idempotency_key: { type: string }
      required: [idempotency_key]
    EncryptionKeyRevokeRequest:
      type: object
      properties:
        idempotency_key: { type: string }
        reason:          { type: string }
      required: [idempotency_key]
    EncryptionKeyRotationResponse:
      type: object
      properties:
        rotation_id:     { type: string }
        tenant_id:       { type: string }
        previous_key_id: { type: string }
        rotated_key:
          $ref: "#/components/schemas/EncryptionKey"
        rotated_at: { type: string, format: date-time }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
    EncryptionSealRequest:
      type: object
      properties:
        idempotency_key: { type: string }
        key_id:          { type: string }
        object_uri:      { type: string }
      required: [idempotency_key, key_id, object_uri]
    EncryptionSealResponse:
      type: object
      properties:
        key_id:            { type: string }
        tenant_id:         { type: string }
        object_uri:        { type: string }
        sealed_object_uri: { type: string }
        unseal_token:      { type: string }
        expires_at:        { type: string, format: date-time }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
    EncryptionUnsealTokenRequest:
      type: object
      properties:
        key_id:            { type: string }
        sealed_object_uri: { type: string }
      required: [key_id, sealed_object_uri]
    EncryptionUnsealTokenResponse:
      type: object
      properties:
        key_id:            { type: string }
        tenant_id:         { type: string }
        sealed_object_uri: { type: string }
        unseal_token:      { type: string }
        expires_at:        { type: string, format: date-time }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
...
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code:        { type: string, example: "MODEL_NOT_FOUND" }
        message:     { type: string, example: "模型 qwen2.5-72b 不存在" }
        request_id:  { type: string, example: "req_01j8xabc" }
        details:     { type: object, additionalProperties: true }

    # Cursor 分页（所有列表接口统一使用此格式）
...
    CoreDevProfileInfo:
      type: object
      description: "Core dev/local profile 标记；用于区分本地联调成功与真实 provider 执行成功。"
      required: [mode, provider, real_provider]
      properties:
        mode:          { type: string, enum: [local, real] }
        provider:      { type: string }
        real_provider: { type: boolean }
        reason:        { type: string, nullable: true }
...
  responses:
...
    NotFound:
      description: 资源不存在（code=NOT_FOUND）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Conflict:
      description: 资源已存在或冲突（code=CONFLICT）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ── API 路径定义 ──────────────────────────────────────────────────────────────
# 说明：/demo/* 路径已迁移至 api/openapi/dev.yaml（不参与 SDK 生成）
paths:
...
  /encryption/keys:
    get:
      operationId: listEncryptionKeys
      summary: 列出加密密钥
      tags: [Encryption]
      x-ani-rbac-scope: "scope:encryption:read"
      parameters:
        - { name: limit,  in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncryptionKeyListResponse"
    post:
      operationId: createEncryptionKey
      summary: 创建加密密钥
      tags: [Encryption]
      x-ani-rbac-scope: "scope:encryption:create"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EncryptionKeyCreateRequest"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncryptionKey"
  /encryption/keys/{key_id}:
    get:
      operationId: getEncryptionKey
      summary: 查询加密密钥
      tags: [Encryption]
      x-ani-rbac-scope: "scope:encryption:read"
      parameters:
        - { name: key_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncryptionKey"
    delete:
      operationId: deleteEncryptionKey
      summary: 删除加密密钥
      tags: [Encryption]
      x-ani-rbac-scope: "scope:encryption:delete"
      parameters:
        - { name: key_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncryptionKey"
  /encryption/keys/{key_id}/rotate:
    post:
      operationId: rotateEncryptionKey
      summary: 轮换加密密钥
      description: Sprint 5 local dev profile 返回新 active key，并将旧 key 标记为 rotated；真实 SM4/KMS provider 接入后必须保持响应语义兼容。
      tags: [Encryption]
      x-ani-rbac-scope: "scope:encryption:manage"
      parameters:
        - { name: key_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EncryptionKeyRotateRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncryptionKeyRotationResponse"
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
  /encryption/keys/{key_id}/revoke:
    post:
      operationId: revokeEncryptionKey
      summary: 吊销加密密钥
      description: Sprint 5 local dev profile 将 key 标记为 revoked；revoked key 不再允许 seal 或生成 unseal token。
      tags: [Encryption]
      x-ani-rbac-scope: "scope:encryption:manage"
      parameters:
        - { name: key_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EncryptionKeyRevokeRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncryptionKey"
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
  /encryption/seal:
    post:
      operationId: sealEncryptionObject
      summary: 加密对象路径并生成解封令牌
      description: Sprint 5 local dev profile 返回模拟 sealed object URI 和 unseal token；真实 SM4/KMS provider 接入后必须保持响应语义兼容。
      tags: [Encryption]
      x-ani-rbac-scope: "scope:encryption:use"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EncryptionSealRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncryptionSealResponse"
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
  /encryption/unseal-token:
    post:
      operationId: createEncryptionUnsealToken
      summary: 生成对象解封令牌
      description: 为 Init Container 或后续模型加载流程生成限时解封令牌；当前为 local dev profile 模拟令牌。
      tags: [Encryption]
      x-ani-rbac-scope: "scope:encryption:use"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EncryptionUnsealTokenRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncryptionUnsealTokenResponse"
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
...